# openapi.yaml
openapi: 3.0.3
info:
  title: LifePlace API
  version: 1.0.0
  description: >
    Comprehensive API documentation for LifePlace Event Management Platform.
    Built with Django REST Framework following Domain-Driven Design principles.
    This CRM system manages the complete event lifecycle from booking flows 
    to payment processing, contracts, and client communications.
  contact:
    name: LifePlace API Support
    url: https://api.crm.example.com/support
  license:
    name: Proprietary
    url: https://api.crm.example.com/license
servers:
  - url: https://api.crm.example.com/v1
    description: Production server
  - url: https://staging-api.crm.example.com/v1
    description: Staging server
  - url: http://localhost:8000
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/questionnaires/:
    get:
      tags: [Questionnaires]
      summary: List questionnaires
      description: Get all questionnaires
      operationId: list_questionnaires
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Questionnaires retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedQuestionnaireList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Questionnaires]
      summary: Create questionnaire
      description: Create a new questionnaire
      operationId: create_questionnaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Questionnaire'
      responses:
        '201':
          description: Questionnaire created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Questionnaire'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/questionnaires/{id}/:
    get:
      tags: [Questionnaires]
      summary: Get questionnaire
      description: Get questionnaire details
      operationId: get_questionnaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Questionnaire retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Questionnaire'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Questionnaires]
      summary: Update questionnaire
      description: Update questionnaire
      operationId: update_questionnaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Questionnaire'
      responses:
        '200':
          description: Questionnaire updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Questionnaire'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Questionnaires]
      summary: Delete questionnaire
      description: Delete questionnaire
      operationId: delete_questionnaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Questionnaire deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/notes/:
    get:
      tags: [Notes]
      summary: List notes
      description: Get notes with filtering by entity
      operationId: list_notes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - name: entity_type
          in: query
          schema:
            type: string
            enum: [event, client, contract, payment]
        - name: entity_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Notes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNoteList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Notes]
      summary: Create note
      description: Create a new note
      operationId: create_note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        '201':
          description: Note created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/notes/{id}/:
    get:
      tags: [Notes]
      summary: Get note
      description: Get note details
      operationId: get_note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Note retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Notes]
      summary: Update note
      description: Update note content
      operationId: update_note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        '200':
          description: Note updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Notes]
      summary: Delete note
      description: Delete a note
      operationId: delete_note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Note deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Notifications Domain Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        default: 1
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string

    # User Management Schemas
...
    Questionnaire:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        description:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/QuestionnaireField'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - name
    QuestionnaireField:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        questionnaire:
          type: integer
        label:
          type: string
        field_type:
          type: string
          enum: [text, textarea, select, multiselect, checkbox, radio, date, number, email, phone]
        options:
          type: array
          items:
            type: string
        required:
          type: boolean
        order:
          type: integer
        validation:
          type: object
      required:
        - questionnaire
        - label
        - field_type
...
    Note:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        entity_type:
          type: string
          enum: [event, client, contract, payment]
        entity_id:
          type: integer
        content:
          type: string
        author:
          type: integer
        is_pinned:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - entity_type
        - entity_id
        - content

    # Notifications Schemas
...
    PaginatedQuestionnaireList:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/Questionnaire'
...
    PaginatedNoteList:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/Note'
...